fix(codegen): DEPYLER-1135 Remove unnecessary numeric casts in push operations
The generate_typed_push function was unconditionally casting integers to i64
and floats to f64. This caused E0308 type mismatch errors when pushing to
Vec<i32> because Python's `int` maps to i32 by default, not i64.

Fix:
- Removed `as i64` cast for Type::Int - trust the type system
- Removed `as f64` cast for Type::Float - trust the type system
- If there's a type mismatch, it should be fixed at the source (type inference)
  not papered over with casts

Also fixes missing struct fields (module_aliases, alias) in test files that
were added in DEPYLER-1136-1139 but not updated in all test locations.

Results:
- Filtered corpus (non-test files): 32.6% → 37.2% (+2 files compiling)
- All 11,309 tests pass
- functional_programming_combined.py now compiles (was failing with E0308)

+++ examples/data_structures/queue.rs
@@ -2724,7 +2724,7 @@ pub fn level_order_values(
             .cloned()
             .expect("IndexError: list index out of range");
         if value.is_some() {
-            result.push(value as i64);
+            result.push(value);
             let left_child = ((2).py_mul(index)).py_add(1);
             let right_child = ((2).py_mul(index)).py_add(2);
             if left_child < tree_values.len() as i32 {
